+++ inst/plumber/05-static/plumber.R
@@ -5,7 +5,7 @@ list()
 list()
 
 #* @get /
-#* @json(auto_unbox = TRUE)
+#* @serializer json list(auto_unbox = TRUE)
 function() {
   "static file server at './files'"
 }